Consolidate semantic domain types on the frontend
