+++ Backend/Models/SemanticDomain.cs
@@ -163,4 +163,27 @@ namespace BackendFramework.Models
             return HashCode.Combine(Node, Parent, Left, Right, Children);
         }
     }
+
+    /// <remarks>
+    /// This is used in an OpenAPI return value serializer, so its attributes must be defined as properties.
+    /// </remarks>
+    public class SemanticDomainWithSubdomains
+    {
+        [Required]
+        public string Name { get; set; }
+        [Required]
+        public string Id { get; set; }
+        [Required]
+        public string Description { get; set; }
+        [Required]
+        public List<SemanticDomainWithSubdomains> Subdomains { get; set; }
+
+        public SemanticDomainWithSubdomains(SemanticDomain sd)
+        {
+            Name = sd.Name;
+            Id = sd.Id;
+            Description = "";
+            Subdomains = new List<SemanticDomainWithSubdomains>();
+        }
+    }
 }
