+++ src/components/DataEntry/DataEntryHeader/tests/DataEntryHeader.test.tsx
@@ -7,7 +7,7 @@ import "tests/mockReactI18next";
 import DataEntryHeader, {
   getQuestions,
 } from "components/DataEntry/DataEntryHeader/DataEntryHeader";
-import TreeSemanticDomain from "components/TreeView/TreeSemanticDomain";
+import { TreeSemanticDomain } from "types/semanticDomain";
 
 const mockStore = configureMockStore()();
 const mockCallback = jest.fn();
