+++ src/components/DataEntry/DataEntryTable/NewEntry/tests/NewEntry.test.tsx
@@ -3,7 +3,7 @@ import renderer from "react-test-renderer";
 import "tests/mockReactI18next";
 
 import NewEntry from "components/DataEntry/DataEntryTable/NewEntry/NewEntry";
-import { newSemanticDomain } from "types/word";
+import { newSemanticDomain } from "types/semanticDomain";
 import { newWritingSystem } from "types/writingSystem";
 
 jest.mock("components/Pronunciations/PronunciationsComponent", () => "div");
