+++ src/components/DataEntry/ExistingDataTable/tests/ExistingDataTable.test.tsx
@@ -3,7 +3,7 @@ import renderer from "react-test-renderer";
 import configureMockStore from "redux-mock-store";
 
 import { ExistingDataTable } from "components/DataEntry/ExistingDataTable/ExistingDataTable";
-import { newSemanticDomain } from "types/word";
+import { newSemanticDomain } from "types/semanticDomain";
 
 jest.mock("components/DataEntry/ExistingDataTable/ImmutableExistingData");
