+++ src/components/TreeView/tests/TreeViewReducer.test.tsx
@@ -1,4 +1,3 @@
-import TreeSemanticDomain from "components/TreeView/TreeSemanticDomain";
 import {
   TreeViewAction,
   TreeActionType,
@@ -9,6 +8,7 @@ import {
   TreeViewState,
 } from "components/TreeView/TreeViewReducer";
 import { StoreAction, StoreActionTypes } from "rootActions";
+import { TreeSemanticDomain } from "types/semanticDomain";
 
 describe("Test the TreeViewReducer", () => {
   it("Returns defaultState when passed undefined", () => {
