+++ src/goals/ReviewEntries/ReviewEntriesComponent/tests/MockWords.ts
@@ -3,13 +3,8 @@ import {
   ReviewEntriesSense,
   ReviewEntriesWord,
 } from "goals/ReviewEntries/ReviewEntriesComponent/ReviewEntriesTypes";
-import {
-  newFlag,
-  newNote,
-  newSemanticDomain,
-  newSense,
-  newWord,
-} from "types/word";
+import { newSemanticDomain } from "types/semanticDomain";
+import { newFlag, newNote, newSense, newWord } from "types/word";
 import { Bcp47Code } from "types/writingSystem";
 
 export default function mockWords(): ReviewEntriesWord[] {
